renaming level-1 package name to one
