+++ openshift/capabilities-tool-job.yaml
@@ -7,8 +7,8 @@ spec:
     spec:
       containers:
       - name: opcap
-        image: quay.io/opdev/opcap:v4.9
-        args: ["level-1","--container-engine","podman","--output-path","/opt/capabilities-tool","--bundle-image","$(bundle_image)","--bucket-name","$(bucket_name)","--bundle-name","$(bundle_name)"]
+        image: quay.io/opdev/opcap:v0.0.1
+        args: ["one","--container-engine","podman","--output-path","/opt/capabilities-tool","--bundle-image","$(bundle_image)","--bucket-name","$(bucket_name)","--bundle-name","$(bundle_name)"]
         env:
           - name: bucket_name
             valueFrom:
