+++ pkg/actions/extract_index.go
@@ -23,7 +23,7 @@ import (
 	"opcap/pkg"
 )
 
-const catalogIndex = "audit-catalog-level-1"
+const catalogIndex = "audit-catalog-one"
 
 func ExtractIndexDB(image string, containerEngine string) error {
 	log.Info("Extracting database...")
@@ -39,10 +39,10 @@ func ExtractIndexDB(image string, containerEngine string) error {
 	}
 
 	// Extract
-	command = exec.Command(containerEngine, "cp", fmt.Sprintf("%s:/database/level-1.db", catalogIndex), "./output/")
+	command = exec.Command(containerEngine, "cp", fmt.Sprintf("%s:/database/one.db", catalogIndex), "./output/")
 	_, err = pkg.RunCommand(command)
 	if err != nil {
-		return fmt.Errorf("unable to extract the image for level-1.db %s : %s", image, err)
+		return fmt.Errorf("unable to extract the image for one.db %s : %s", image, err)
 	}
 	return nil
 }
